+++ RELEASE-NOTES.md
@@ -51,6 +51,7 @@
 - **Menu** - Fix text align on `dropdown item` inside `icon menu`
 - **Menu** - Fixed issue with `labeled input` text inside menu not appearing vertically centered
 - **Menu** - Fixed hybrid initialization not creating `menu` correctly. Fixed docs example of hybrid `<select>` initialization
+- **Dropdown/Tab** - Fixed an instance where `metadata` was not referencing settings metadata value
 
 **[Merged PR](https://github.com/Semantic-Org/Semantic-UI/issues?q=is%3Aissue+milestone%3A2.1.0+is%3Aclosed)**